requirements.txt
# =====================
# Agent Graph RAG - 依赖清单
# =====================

# 核心依赖
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0

# 嵌入模型
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0
torch>=2.0.0
transformers>=4.30.0

# 向量数据库
chromadb>=0.4.0

# LLM 模型
openai>=1.0.0

# 数据处理
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0

# OSS 存储
oss2>=2.18.0

# PDF 处理
pymupdf>=1.23.0

# 数据解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 其他工具
tqdm>=4.66.0
requests>=2.31.0
